+++ .markdownlint.json
@@ -7,7 +7,8 @@
       "details",
       "summary",
       "div",
-      "img"
+      "img",
+      "br"
     ]
   },
   "MD013": false,
