+++ docusaurus/docs/develop/path/4_cheatsheet_morse.md
@@ -123,7 +123,7 @@ Run the following commands to generate a Morse config at `local/path/.config.yam
 make morse_prepare_e2e_config
 
 # Copy to ./local/path/.config.yaml
-make morse_copy_e2e_config_to_local
+make morse_copy_e2e_load_test_config_to_local
 
 # IMPORTANT: In the next, step, update ./local/path/.config.yaml
 ```
