+++ e2e/config/e2e_load_test.config.tmpl.yaml
@@ -1,0 +1,133 @@
+# yaml-language-server: $schema=./e2e_load_test.config.schema.yaml
+
+# Default Load test configuration for Morse and Shannon
+#
+# Usage:
+# - Expected config location: `.e2e_load_test.config.yaml`
+# - If no custom config is found, the default template is used.
+# - Create your own config to override defaults.
+#
+# Vegeta Test modes:
+# 1. E2E Test Mode: Starts PATH in Docker (Dockertest) & runs tests against it
+# 2. Load Test Mode: Runs tests against a specified gateway URL
+
+# General test configuration overrides for all tests
+e2e_load_test_config:
+  # E2E Test Mode
+  # PATH runs in Docker (Dockertest), tests run against this container
+  e2e_config:
+    # [Optional] Seconds to wait for hydrator checks after PATH startup
+    wait_for_hydrator: 0
+    # [Optional] Docker-related options
+    docker_config:
+      # [Optional] Log Docker container output to a file
+      log_to_file: false
+      # [Optional] Force Docker image rebuild (useful after code changes)
+      force_rebuild_image: false
+
+  # Load Test Mode
+  # Tests run against a specified gateway URL (local or public)
+  # Examples:
+  #   - Local:  "http://localhost:3091/v1"
+  #   - Public: "https://rpc.grove.city/v1"
+  load_test_config:
+    # [Required] Custom PATH gateway URL
+    gateway_url_override: ""
+    # [Required] Custom user identifier for the test (i.e. portal-application-id)
+    portal_application_id: ""
+    # [Optional] Insert service ID into subdomain per test case
+    # - Example: "F00C.rpc.grove.city/v1"
+    # - TODO_TECHDEBT(@commoddity): Remove after header-based service selection is supported
+    # - See: https://github.com/buildwithgrove/infrastructure/issues/91
+    use_service_subdomain: false
+    # [Optional] Custom API key for the test (i.e. portal-api-key)
+    portal_api_key: ""
+
+# Default configuration applied to all test cases (unless overridden)
+default_test_config:
+  global_rps: 50 # Requests per second (shared across all methods)
+  requests_per_method: 50 # Number of requests per method
+  success_rate: 0.80 # Minimum required success rate (80%)
+  max_p50_latency_ms: 2000ms # Max allowed P50 latency (ms)
+  max_p95_latency_ms: 5000ms # Max allowed P95 latency (ms)
+  max_p99_latency_ms: 10000ms # Max allowed P99 latency (ms)
+
+test_cases:
+  ###########################################
+  # ---------------- MORSE ----------------
+  ###########################################
+
+  # Ethereum Mainnet (Archival)
+  - name: "F00C (Ethereum) Load Test"
+    protocol: "morse"
+    service_id: "F00C"
+    archival: true
+    service_params:
+      contract_address: "0x28C6c06298d514Db089934071355E5743bf21d60"
+      contract_start_block: 12300000 # Minimum block for archival tests
+      transaction_hash: "0xfeccd627b5b391d04fe45055873de3b2c0b4302d52e96bd41d5f0019a704165f"
+      call_data: "0x18160ddd"
+
+  # Polygon Mainnet (Archival)
+  - name: "F021 (Polygon) Load Test"
+    protocol: "morse"
+    service_id: "F021"
+    archival: true
+    service_params:
+      contract_address: "0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270"
+      contract_start_block: 5000000
+      transaction_hash: "0xb4f33e8516656d513df5d827323003c7ad1dcbb5bc46dff57c9bebad676fefe4"
+      call_data: "0x18160ddd"
+
+  # Oasys Mainnet (Archival)
+  - name: "F01C (Oasys) Load Test"
+    protocol: "morse"
+    service_id: "F01C"
+    archival: true
+    service_params:
+      contract_address: "0xf89d7b9c864f589bbF53a82105107622B35EaA40"
+      contract_start_block: 424300
+      transaction_hash: "0x7e5904f6f566577718aa3ddfe589bb6d553daaeb183e2bdc63f5bf838fede8ee"
+      call_data: "0x18160ddd"
+
+  # XRPL EVM Testnet (Archival)
+  - name: "F036 (XRPL EVM Testnet) Load Test"
+    protocol: "morse"
+    service_id: "F036"
+    archival: true
+    service_params:
+      contract_address: "0xc29e2583eD5C77df8792067989Baf9E4CCD4D7fc"
+      contract_start_block: 368266
+      transaction_hash: "0xa59fde70cac38068dfd87adb1d7eb40200421ebf7075911f83bcdde810e94058"
+      call_data: "0x18160ddd"
+
+  ###########################################
+  # ---------------- SHANNON ----------------
+  ###########################################
+
+  # Anvil local Ethereum node (Ephemeral)
+  - name: "anvil (local Ethereum) Load Test"
+    protocol: "shannon"
+    service_id: "anvil"
+    service_params:
+      contract_address: "0xdAC17F958D2ee523a2206206994597C13D831ec7"
+      call_data: "0x18160ddd"
+    # Multiply latency thresholds by this value (for slower networks)
+    latency_multiplier: 10
+    # Override default config for this Shannon service test case
+    test_case_config_override:
+      global_rps: 1
+      requests_per_method: 3
+      success_rate: 0.60
+      max_p50_latency_ms: 30000ms
+      max_p95_latency_ms: 60000ms
+      max_p99_latency_ms: 90000ms
+    # Only test these EVM JSON-RPC methods for this Shannon test case
+    test_case_method_override:
+      - eth_blockNumber
+      - eth_call
+      - eth_getBlockByNumber
+      - eth_getBalance
+      - eth_chainId
+      - eth_getTransactionCount
+      - eth_gasPrice
