+++ makefiles/configs.mk
@@ -9,6 +9,7 @@ CONFIG_FILES := \
 	./local/path/.config.yaml \
 	./e2e/config/.shannon.config.yaml \
 	./e2e/config/.morse.config.yaml \
+	./e2e/config/.e2e_load_test.config.yaml \
 	./local/path/envoy/.envoy.yaml
 
 # Helper function to check if a config file exists
